Properly load and unload proxy lib
